requirements.txt
asgiref==3.4.1
Django==3.2.6
django-filter==2.4.0
django-session-timeout==0.1.0
django-widget-tweaks==1.4.8
docopt==0.6.2
num2words==0.5.10
pkg-resources==0.0.0
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
